Drop load % gauge, add pulse image for future animation.
The percentage load gauge shows no more info than the load gauge, and the
amount is already implied by the scale.

The pulse image will be animated so that incoming and outgoing charge is shown
as a moving pulse. This commit just gets the image in, and the CSS right.